Split transport into its own file, install node packages
