+++ package.json
@@ -25,8 +25,12 @@
     "express": "^4.13.3",
     "joi": "^10.2.2",
     "juice": "^4.0.2",
-    "nodemailer": "2.7.2",
-    "nodemailer-ses-transport": "^1.5.0",
+    "nodemailer": "^4.0.1",
+    "nodemailer-mailgunapi-transport": "0.0.3",
+    "nodemailer-mandrill-transport": "^1.2.0",
+    "nodemailer-postmark-transport": "^1.2.0",
+    "nodemailer-sendgrid-transport": "^0.2.0",
+    "nodemailer-ses-transport": "^1.5.1",
     "nodemailer-stub-transport": "^1.1.0",
     "pug": "^2.0.0-beta11",
     "winston": "*"
